+++ examples/nnfft/accuracy.c
@@ -36,7 +36,7 @@ void accuracy(NFFT_INT d)
   NFFT_INT M_total,N_total;
   M_total=10000;N_total=1;
 
-  slow=(double _Complex*)nfft_malloc(M_total*sizeof(double _Complex));
+  slow=(double _Complex*)nfft_malloc((size_t)M_total * sizeof(double _Complex));
 
   for(t=0; t<d; t++)
     {
